Project restructure. - All versions of the VPN now run in the same container.
