--- X25519/docker-compose.yml
@@ -1,82 +1,0 @@
-services:
-    VPN_Client:
-        build:
-          context: .
-          dockerfile: Dockerfile
-        image: handsonsecurity/seed-ubuntu:large
-        container_name: X25519-client-10.9.0.5
-        tty: true
-        cap_add:
-                - ALL
-        devices:
-                - "/dev/net/tun:/dev/net/tun"
-        volumes:
-                - ./volumes:/volumes
-                - ./keys:/keys
-
-        networks:
-            net-10.9.0.0:
-                ipv4_address: 10.9.0.5
-        command: bash -c "
-                     tail -f /dev/null
-                 "
-
-    Host1:
-        build:
-          context: .
-          dockerfile: Dockerfile
-        image: handsonsecurity/seed-ubuntu:large
-        container_name: X25519-host-192.168.60.7
-        tty: true
-        cap_add:
-                - ALL
-        networks:
-            net-192.168.60.0:
-                ipv4_address: 192.168.60.7
-        command: bash -c "
-                      ip route del default  &&
-                      ip route add default via 192.168.60.11  &&
-                      /etc/init.d/openbsd-inetd start &&
-                      tail -f /dev/null
-                 "
-
-    Router:
-        build:
-          context: .
-          dockerfile: Dockerfile
-        image: handsonsecurity/seed-ubuntu:large
-        container_name: X25519-server-router
-        tty: true
-        cap_add:
-                - ALL
-        devices:
-                - "/dev/net/tun:/dev/net/tun"
-        sysctls:
-                - net.ipv4.ip_forward=1
-        volumes:
-          - ./volumes:/volumes
-          - ./keys:/keys
-        networks:
-            net-10.9.0.0:
-                ipv4_address: 10.9.0.11
-            net-192.168.60.0:
-                ipv4_address: 192.168.60.11
-        command: bash -c "
-                      ip route del default  &&
-                      ip route add default via 10.9.0.1 &&
-                      tail -f /dev/null
-                 "
-
-networks:
-    net-192.168.60.0:
-        name: net-192.168.60.0
-        ipam:
-            config:
-                - subnet: 192.168.60.0/24
-
-    net-10.9.0.0:
-        name: net-10.9.0.0
-        ipam:
-            config:
-                - subnet: 10.9.0.0/24
-
