--- X25519/requirements.txt
@@ -1,2 +1,0 @@
-scapy~=2.6.1
-cryptography~=44.0.2
