--- X25519/volumes/shared/crypto/decrypt.py
@@ -1,20 +1,0 @@
-from cryptography.hazmat.primitives.asymmetric import x25519
-from cryptography.hazmat.primitives.kdf.hkdf import HKDF
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives import serialization, hashes
-from cryptography.hazmat.backends import default_backend
-from shared.crypto.tools import derive_shared_key, NONCE_SIZE
-import os
-
-# AES-256-GCM
-def decrypt(ciphertext, private_key, peer_public_key):
-    key = derive_shared_key(private_key, peer_public_key)
-    nonce = ciphertext[:NONCE_SIZE]
-    tag = ciphertext[NONCE_SIZE:NONCE_SIZE+16]
-    ct = ciphertext[NONCE_SIZE+16:]
-    decryptor = Cipher(
-        algorithms.AES(key),
-        modes.GCM(nonce, tag),
-        backend=default_backend()
-    ).decryptor()
-    return decryptor.update(ct) + decryptor.finalize()
