--- X25519/volumes/shared/crypto/encrypt.py
@@ -1,19 +1,0 @@
-from cryptography.hazmat.primitives.asymmetric import x25519
-from cryptography.hazmat.primitives.kdf.hkdf import HKDF
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives import serialization, hashes
-from cryptography.hazmat.backends import default_backend
-from shared.crypto.tools import derive_shared_key, NONCE_SIZE
-import os
-
-# AES-256-GCM
-def encrypt(plaintext, private_key, peer_public_key):
-    key = derive_shared_key(private_key, peer_public_key)
-    nonce = os.urandom(NONCE_SIZE)
-    encryptor = Cipher(
-        algorithms.AES(key),
-        modes.GCM(nonce),
-        backend=default_backend()
-    ).encryptor()
-    ciphertext = encryptor.update(plaintext) + encryptor.finalize()
-    return nonce + encryptor.tag + ciphertext
